// src/model/FireLog/FireLogDetailModel.ts
export interface FireLogDetailModel {
    id: number;
    fireSource: string;
    fullNameCustomer: string;
    constructionName: string;
    constructionFullAddress: string;
    username: string;
    status: string;
    dateCreated: number;
    dateSolved: number;
    dateModified: number;
    longitude: number;
    latitude: number;
    provinceId: number;
    districtId: number;
}